refactor: WalletModal from OptionModal
